+++ capture_gui/app.py
@@ -320,7 +320,7 @@ class App(QtWidgets.QWidget):
         title_version = "{} v{}".format(title, version.version)
         self.setObjectName(self.object_name)
         self.setWindowTitle(title_version)
-        self.setMinimumWidth(400)
+        self.setMinimumWidth(440)
 
         # Set dialog window flags so the widget can be correctly parented
         # to Maya main window
@@ -374,8 +374,9 @@ class App(QtWidgets.QWidget):
             return
 
         options = self.get_outputs()
-        if options["frame"] is not None:
-            options["raw_frame_numbers"] = True
+        frames = options.get("frame", None)
+        if frames is None:
+            options["raw_frame_numbers"] = False
         filename = options.get("filename", None)
 
         self.playblast_start.emit(options)